requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
httpx
msal
